# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ARN INFRA API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  operator: "operator",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  createdAt: string;
}

export interface LoginResponse {
  token: string;
  user: User;
}

export type CreateUserBodyRole =
  (typeof CreateUserBodyRole)[keyof typeof CreateUserBodyRole];

export const CreateUserBodyRole = {
  admin: "admin",
  operator: "operator",
} as const;

export interface CreateUserBody {
  name: string;
  email: string;
  password: string;
  role: CreateUserBodyRole;
}

export type UpdateUserBodyRole =
  (typeof UpdateUserBodyRole)[keyof typeof UpdateUserBodyRole];

export const UpdateUserBodyRole = {
  admin: "admin",
  operator: "operator",
} as const;

export interface UpdateUserBody {
  name?: string;
  email?: string;
  role?: UpdateUserBodyRole;
  password?: string;
}

export interface WorkEntry {
  id: number;
  date: string;
  labourCount: number;
  squareMeter: number;
  workingChannel: string;
  createdBy: number;
  createdByName?: string;
  createdAt: string;
}

export interface CreateWorkEntryBody {
  date: string;
  labourCount: number;
  squareMeter: number;
  workingChannel: string;
}

export interface UpdateWorkEntryBody {
  date?: string;
  labourCount?: number;
  squareMeter?: number;
  workingChannel?: string;
}

export interface ReportSummary {
  totalEntries: number;
  totalLabour: number;
  totalSquareMeter: number;
  entries: WorkEntry[];
}

export interface DashboardStats {
  todayLabour: number;
  todaySquareMeter: number;
  monthLabour: number;
  monthSquareMeter: number;
  totalEntries: number;
  recentEntries: WorkEntry[];
}

export type GetWorkEntriesParams = {
  startDate?: string;
  endDate?: string;
};

export type GetReportSummaryParams = {
  startDate?: string;
  endDate?: string;
};

export type ExportExcelParams = {
  startDate?: string;
  endDate?: string;
};

export type ExportPdfParams = {
  startDate?: string;
  endDate?: string;
};