# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ARN INFRA API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WorkEntry } from "./workEntry";

export interface DashboardStats {
  todayLabour: number;
  todaySquareMeter: number;
  monthLabour: number;
  monthSquareMeter: number;
  totalEntries: number;
  recentEntries: WorkEntry[];
}