# lib/api-zod/src/generated/types/reportSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ARN INFRA API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WorkEntry } from "./workEntry";

export interface ReportSummary {
  totalEntries: number;
  totalLabour: number;
  totalSquareMeter: number;
  entries: WorkEntry[];
}